requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
Flask==0.12.2
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
kafka-python==1.4.7
MarkupSafe==1.0
mmh3==2.5.1
python-snappy==0.5.4
requests==2.18.4
urllib3==1.22
Werkzeug==0.14.1
